port xml utils from future
