+++ kratos/utilities/xml_utilities/xml_ascii_nd_data_element.h
@@ -20,7 +20,7 @@
 #include "containers/nd_data.h"
 #include "xml_nd_data_element.h"
 
-namespace Kratos::Future {
+namespace Kratos {
 
 ///@name Kratos Classes
 ///@{
@@ -67,4 +67,4 @@ private:
 
 ///@}
 
-} // namespace Kratos::Future
+} // namespace Kratos
