+++ kratos/utilities/xml_utilities/xml_base64_binary_nd_data_element.h
@@ -20,7 +20,7 @@
 #include "containers/nd_data.h"
 #include "xml_nd_data_element.h"
 
-namespace Kratos::Future {
+namespace Kratos {
 
 ///@name Kratos Classes
 ///@{
@@ -59,4 +59,4 @@ public:
 
 ///@}
 
-} // namespace Kratos::Future
+} // namespace Kratos
