+++ kratos/utilities/xml_utilities/xml_data_element_wrapper.h
@@ -20,7 +20,7 @@
 #include "containers/nd_data.h"
 #include "xml_element.h"
 
-namespace Kratos::Future {
+namespace Kratos {
 
 ///@name Kratos Classes
 ///@{
@@ -64,4 +64,4 @@ public:
     ///@}
 };
 
-} // namespace Kratos::Future
+} // namespace Kratos
